Modify code to fit uint16 requirements.

+++ training/dataset.py
@@ -21,7 +21,7 @@ def parse_tfrecord_tf(record):
     features = tf.parse_single_example(record, features={
         'shape': tf.FixedLenFeature([3], tf.int64),
         'data': tf.FixedLenFeature([], tf.string)})
-    data = tf.decode_raw(features['data'], tf.uint8)
+    data = tf.decode_raw(features['data'], tf.uint16)
     return tf.reshape(data, features['shape'])
 
 def parse_tfrecord_np(record):
@@ -29,7 +29,7 @@ def parse_tfrecord_np(record):
     ex.ParseFromString(record)
     shape = ex.features.feature['shape'].int64_list.value # temporary pylint workaround # pylint: disable=no-member
     data = ex.features.feature['data'].bytes_list.value[0] # temporary pylint workaround # pylint: disable=no-member
-    return np.fromstring(data, np.uint8).reshape(shape)
+    return np.fromstring(data, np.uint16).reshape(shape)
 
 #----------------------------------------------------------------------------
 # Dataset class that loads data from tfrecords files.
@@ -50,7 +50,7 @@ class TFRecordDataset:
         self.resolution         = None
         self.resolution_log2    = None
         self.shape              = []        # [channel, height, width]
-        self.dtype              = 'uint8'
+        self.dtype              = 'uint16'
         self.dynamic_range      = [0, 26350]
         self.label_file         = label_file
         self.label_size         = None      # [component]
@@ -169,7 +169,7 @@ class TFRecordDataset:
 # Base class for datasets that are generated on the fly.
 
 class SyntheticDataset:
-    def __init__(self, resolution=1024, num_channels=3, dtype='uint8', dynamic_range=[0,26350], label_size=0, label_dtype='float32'):
+    def __init__(self, resolution=1024, num_channels=3, dtype='uint16', dynamic_range=[0,26350], label_size=0, label_dtype='float32'):
         self.resolution         = resolution
         self.resolution_log2    = int(np.log2(resolution))
         self.shape              = [num_channels, resolution, resolution]
